requirements.txt
discord.py
PyYAML
mongoengine
dice